+++ CHANGELOG.markdown
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 * Changed the [Expression Optimization](doc/syntax/optimizations/EXPRESSION-OPTIMIZATION.markdown) to handle integer division by one the same as the `op floor` instruction.
 * Changed the mlog parser to raise an error when the code contains duplicate labels, regardless of the processor version. This behavior matches the latest Mindustry version; in older Mindustry versions the code is still accepted. 
 * Changed the [Expression Optimization](doc/syntax/optimizations/EXPRESSION-OPTIMIZATION.markdown) and [Array Optimization](doc/syntax/optimizations/ARRAY-OPTIMIZATION.markdown) to [remove an `op floor` instruction when applied to the array index](doc/syntax/optimizations/EXPRESSION-OPTIMIZATION.markdown#handling-non-negative-implicit-integer-conversions).
+* Changed the [Expression Optimization](doc/syntax/optimizations/EXPRESSION-OPTIMIZATION.markdown): when the `op round` instruction is available, a sequence of `op add` and `op floor` instructions (as in `floor(x + 0.5)`) is replaced with the `op round` instruction (as in `round(x)`).
 * The variable reference index is recreated before each optimization, potentially allowing more optimizations to be performed.
 
 ## 3.16.2 – 2026-04-25